ASoC: soc: msm: dynamic change port id for speaker and headset
Since we have quat mi2s for speaker, the port id for srs trumedia need to
change to AFE_PORT_ID_QUATERNARY_MI2S_RX when speaker playback,
so that it can hear the dts effect real-time.

libsrsprocessing.so handles the decision on which "SRS TruMedia *" control
to use when applying the settings.  The right way to do this is to add a
"SRS TruMedia QUAT_MI2S" control, and modify libsrsprocessing.so to choose
the correct mixer control depending on output device.  However, we don't
have access to the proprietary source, and QCOM is unresponsive to the
request.

Change-Id: Iabe7ceb58a628d094ee00a8f4ef967f345961e46