exit: move profile_task_exit() call to after setting PF_EXITING
Move the profiling point to immediately after PF_EXITING is set. For
existing handlers of the PROFILE_TASK_EXIT notifier, this difference
should be inconsequential, but it makes synchronization EASIER in
profilers such as the uid_cputime driver, which may be iterating over
the task list in a context besides a PROFILE_TASK_EXIT notifier.

Change-Id: I0ddccbfb577a0fd1d094fc610b9fb8cb38a49104
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>