uid_cputime: revert changes depending on non-existing task struct field 'cpu_power'
This reverts ee2d868d0c57 ("uid_cputime: Avoids double accounting of
process stime, utime and cpu_power in task exit.") and 6b4efb1aa693
("uid_cputime: Extends the cputime functionality to report power per
uid").

These changes depend on a cpu_power field in the task struct which,
along with the power awareness features that update it, are not
present in this kernel branch.

Revert the changes so that uid_cputime can compile once-again.

Change-Id: Id86ea353e541c26be4fed621d493b3279c560f46
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>